feat: PomEditSession, change tracking, managed deps, unified UpdatesTui
Introduce shared PomEditSession per module with snapshot-based undo, change
log, save with external-modification check, and visual change markers (+/~/-)
on table rows across all editing tools.

Key changes:
- PomEditSession: shared edit state, undo stack, change tracking, safe save
- DependenciesTui: new Managed tab for dependencyManagement/BOM entries
- Unified UpdatesTui: merged ReactorUpdatesTui into UpdatesTui with single-
  module mode (hides Modules tab and "N mod" column when only one module)
- Key bindings: e/w→E/W in read-only tools, s→f for filter, h→? for help,
  freeing w/z/R/d for edit lifecycle
- All editing tools (Dependencies, Updates, Conflicts, Audit) use shared
  PomEditSession with beforeMutation()/recordChange() pattern
- Tab bar shows per-tool change counts [+N ~N -N]
- PilotShell: session cache, centralized quit with unsaved-changes prompt
- Audit: fixed aggregation to use GAV keys (not GA) for correct dedup
