+++ pilot-core/src/test/java/eu/maveniverse/maven/pilot/UpdatesTuiTest.java
@@ -19,14 +19,17 @@
 package eu.maveniverse.maven.pilot;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.assertj.core.api.Assertions.assertThatNoException;
 
+import dev.tamboui.terminal.Terminal;
+import dev.tamboui.terminal.TestBackend;
+import dev.tamboui.tui.event.KeyCode;
+import dev.tamboui.tui.event.KeyEvent;
+import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.concurrent.Executors;
-import org.junit.jupiter.api.BeforeEach;
+import java.util.Properties;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
@@ -35,273 +38,660 @@ class UpdatesTuiTest {
     @TempDir
     Path tempDir;
 
-    String pomPath;
-
-    @BeforeEach
-    void setUp() throws Exception {
-        pomPath = Files.writeString(tempDir.resolve("pom.xml"), "<project/>").toString();
+    private Path subdir(String name) throws IOException {
+        return Files.createDirectories(tempDir.resolve(name));
     }
 
-    @Test
-    void versionResolverInterface() {
-        UpdatesTui.VersionResolver resolver = (g, a) -> List.of("2.0", "1.5", "1.0");
-        List<String> versions = resolver.resolveVersions("com.example", "lib");
-        assertThat(versions).containsExactly("2.0", "1.5", "1.0");
+    private PilotProject createProject(String groupId, String artifactId, String version, Path basedir) {
+        return createProject(groupId, artifactId, version, basedir, List.of(), List.of(), List.of(), List.of());
     }
 
-    @Test
-    void versionResolverThrowingException() {
-        UpdatesTui.VersionResolver resolver = (g, a) -> {
-            throw new IllegalStateException("repo unavailable");
-        };
-        assertThatThrownBy(() -> resolver.resolveVersions("com.example", "lib"))
-                .isInstanceOf(IllegalStateException.class)
-                .hasMessage("repo unavailable");
+    private PilotProject createProject(
+            String groupId,
+            String artifactId,
+            String version,
+            Path basedir,
+            List<PilotProject.Dep> deps,
+            List<PilotProject.Dep> mgmtDeps,
+            List<PilotProject.Dep> origDeps,
+            List<PilotProject.Dep> origMgmtDeps) {
+        return createProject(
+                groupId, artifactId, version, basedir, deps, mgmtDeps, origDeps, origMgmtDeps, new Properties());
     }
 
-    @Test
-    void updateStatusIfDoneNoFailures() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "a", "1.0", "compile", "jar"));
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "b", "2.0", "compile", "jar"));
-        deps.get(0).newestVersion = "1.1";
-        deps.get(0).updateType = VersionComparator.UpdateType.PATCH;
-
-        var tui = new UpdatesTui(deps, pomPath, "com.example:app:1.0", (g, a) -> List.of());
-        tui.loadedCount = 2;
-        tui.updateStatusIfDone();
-
-        assertThat(tui.loading).isFalse();
-        assertThat(tui.status).contains("1 update(s) available").doesNotContain("failed");
+    private PilotProject createProject(
+            String groupId,
+            String artifactId,
+            String version,
+            Path basedir,
+            List<PilotProject.Dep> deps,
+            List<PilotProject.Dep> mgmtDeps,
+            List<PilotProject.Dep> origDeps,
+            List<PilotProject.Dep> origMgmtDeps,
+            Properties origProps) {
+        return new PilotProject(
+                groupId,
+                artifactId,
+                version,
+                "jar",
+                basedir,
+                basedir.resolve("pom.xml"),
+                deps,
+                mgmtDeps,
+                origDeps,
+                origMgmtDeps,
+                origProps,
+                null,
+                null);
     }
 
-    @Test
-    void updateStatusIfDoneWithFailures() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "a", "1.0", "compile", "jar"));
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "b", "2.0", "compile", "jar"));
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "c", "3.0", "compile", "jar"));
-
-        var tui = new UpdatesTui(deps, pomPath, "com.example:app:1.0", (g, a) -> List.of());
-        tui.loadedCount = 3;
-        tui.failedCount = 2;
-        tui.updateStatusIfDone();
-
-        assertThat(tui.loading).isFalse();
-        assertThat(tui.status).contains("0 update(s) available").contains("2 lookup(s) failed");
+    private PilotProject.Dep dep(String groupId, String artifactId, String version) {
+        return new PilotProject.Dep(groupId, artifactId, version);
     }
 
-    @Test
-    void updateStatusIfDoneNotYetComplete() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "a", "1.0", "compile", "jar"));
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "b", "2.0", "compile", "jar"));
-
-        var tui = new UpdatesTui(deps, pomPath, "com.example:app:1.0", (g, a) -> List.of());
-        tui.loadedCount = 1; // not all loaded yet
-        tui.updateStatusIfDone();
-
-        assertThat(tui.loading).isTrue(); // still loading
-        assertThat(tui.status).isEqualTo("Loading updates…"); // unchanged
+    private PilotProject.Dep dep(String groupId, String artifactId, String version, String scope) {
+        return new PilotProject.Dep(groupId, artifactId, version, scope, null);
     }
 
-    @Test
-    void dependencyInfoGa() {
-        var dep = new UpdatesTui.DependencyInfo("org.slf4j", "slf4j-api", "2.0.9", "compile", "jar");
-        assertThat(dep.ga()).isEqualTo("org.slf4j:slf4j-api");
+    private UpdatesTui createTui(ReactorCollector.CollectionResult result, List<PilotProject> projects) {
+        ReactorModel model = ReactorModel.build(projects);
+        return new UpdatesTui(result, model, "com.example:app:1.0", (g, a) -> List.of());
     }
 
-    @Test
-    void dependencyInfoHasUpdate() {
-        var dep = new UpdatesTui.DependencyInfo("com.example", "lib", "1.0", "compile", "jar");
-        assertThat(dep.hasUpdate()).isFalse();
-
-        dep.newestVersion = "1.1";
-        assertThat(dep.hasUpdate()).isTrue();
-
-        dep.newestVersion = "1.0"; // same version
-        assertThat(dep.hasUpdate()).isFalse();
-
-        dep.newestVersion = ""; // empty
-        assertThat(dep.hasUpdate()).isFalse();
-    }
-
-    @Test
-    void dependencyInfoDefaults() {
-        var dep = new UpdatesTui.DependencyInfo("g", "a", null, null, null);
-        assertThat(dep.version).isEmpty();
-        assertThat(dep.scope).isEqualTo("compile");
-        assertThat(dep.type).isEqualTo("jar");
-    }
-
-    @Test
-    void applyVersionResultFindsNewest() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        var dep = new UpdatesTui.DependencyInfo("com.example", "lib", "1.0", "compile", "jar");
-        deps.add(dep);
-
-        var tui = new UpdatesTui(deps, pomPath, "g:a:1.0", (g, a) -> List.of());
-        tui.applyVersionResult(dep, List.of("2.0.0", "1.5.0", "1.0.0"));
-
-        assertThat(dep.newestVersion).isEqualTo("2.0.0");
-        assertThat(dep.updateType).isEqualTo(VersionComparator.UpdateType.MAJOR);
-        assertThat(tui.loadedCount).isEqualTo(1);
-    }
-
-    @Test
-    void applyVersionResultSkipsPreview() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        var dep = new UpdatesTui.DependencyInfo("com.example", "lib", "1.0", "compile", "jar");
-        deps.add(dep);
-
-        var tui = new UpdatesTui(deps, pomPath, "g:a:1.0", (g, a) -> List.of());
-        tui.applyVersionResult(dep, List.of("2.0.0-beta1", "2.0.0-alpha1", "1.1.0"));
-
-        assertThat(dep.newestVersion).isEqualTo("1.1.0");
-        assertThat(dep.updateType).isEqualTo(VersionComparator.UpdateType.MINOR);
-    }
-
-    @Test
-    void applyVersionResultNoUpdate() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        var dep = new UpdatesTui.DependencyInfo("com.example", "lib", "2.0", "compile", "jar");
-        deps.add(dep);
-
-        var tui = new UpdatesTui(deps, pomPath, "g:a:1.0", (g, a) -> List.of());
-        tui.applyVersionResult(dep, List.of("1.5.0", "1.0.0"));
-
-        assertThat(dep.newestVersion).isNull();
-        assertThat(dep.updateType).isNull();
-    }
-
-    @Test
-    void applyVersionResultEmptyVersions() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        var dep = new UpdatesTui.DependencyInfo("com.example", "lib", "1.0", "compile", "jar");
-        deps.add(dep);
-
-        var tui = new UpdatesTui(deps, pomPath, "g:a:1.0", (g, a) -> List.of());
-        tui.applyVersionResult(dep, List.of());
-
-        assertThat(dep.newestVersion).isNull();
-        assertThat(tui.loadedCount).isEqualTo(1);
-    }
-
-    @Test
-    void applyVersionResultWithEmptyCurrentVersion() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        var dep = new UpdatesTui.DependencyInfo("com.example", "lib", null, "compile", "jar");
-        deps.add(dep);
-
-        var tui = new UpdatesTui(deps, pomPath, "g:a:1.0", (g, a) -> List.of());
-        tui.applyVersionResult(dep, List.of("2.0.0", "1.0.0"));
-
-        // Empty version should accept any non-preview version
-        assertThat(dep.newestVersion).isEqualTo("2.0.0");
-    }
-
-    @Test
-    void fetchUpdatesSuccess() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "lib", "1.0", "compile", "jar"));
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "other", "2.0", "compile", "jar"));
-
-        var tui = new UpdatesTui(deps, pomPath, "g:a:1.0", (g, a) -> {
-            if ("lib".equals(a)) return List.of("2.0.0", "1.5.0", "1.0.0");
-            return List.of("2.0.0");
+    private void renderFrame(UpdatesTui tui) {
+        assertThatNoException().isThrownBy(() -> {
+            var terminal = new Terminal<>(new TestBackend(120, 30));
+            terminal.draw(tui::renderStandalone);
         });
+    }
 
-        var executor = Executors.newSingleThreadExecutor();
-        try {
-            tui.fetchUpdates(Runnable::run, executor).join();
-        } finally {
-            executor.shutdownNow();
+    @Test
+    void duplicatePropertyNamesDetected() throws IOException {
+        Path dir1 = subdir("parent");
+        Path dir2 = subdir("child");
+
+        Properties parentProps = new Properties();
+        parentProps.setProperty("lib.version", "1.0");
+        PilotProject parent = createProject(
+                "com.example",
+                "parent-bom",
+                "1.0",
+                dir1,
+                List.of(dep("com.example", "lib-a", "1.0")),
+                List.of(),
+                List.of(dep("com.example", "lib-a", "${lib.version}")),
+                List.of(),
+                parentProps);
+
+        Properties childProps = new Properties();
+        childProps.setProperty("lib.version", "2.0");
+        PilotProject child = createProject(
+                "com.example",
+                "child-bom",
+                "1.0",
+                dir2,
+                List.of(dep("com.example", "lib-b", "2.0")),
+                List.of(),
+                List.of(dep("com.example", "lib-b", "${lib.version}")),
+                List.of(),
+                childProps);
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(parent, child));
+        assertThat(result.propertyGroups).hasSize(2);
+
+        var tui = createTui(result, List.of(parent, child));
+
+        // Set up updates so groups appear in display rows
+        for (var dep : result.allDependencies) {
+            dep.newestVersion = "3.0";
+            dep.updateType = VersionComparator.UpdateType.MAJOR;
+        }
+        for (var group : result.propertyGroups) {
+            group.newestVersion = "3.0";
+            group.updateType = VersionComparator.UpdateType.MAJOR;
         }
 
-        assertThat(deps.get(0).newestVersion).isEqualTo("2.0.0");
-        assertThat(deps.get(0).updateType).isEqualTo(VersionComparator.UpdateType.MAJOR);
-        assertThat(tui.loadedCount).isEqualTo(2);
-        assertThat(tui.loading).isFalse();
-        assertThat(tui.status).contains("update(s) available");
+        tui.buildDisplayRows();
+
+        assertThat(tui.duplicatePropertyNames).containsExactly("lib.version");
     }
 
     @Test
-    void fetchUpdatesWithFailure() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "lib", "1.0", "compile", "jar"));
+    void noDuplicatesWhenPropertyNamesAreUnique() throws IOException {
+        Path dir1 = subdir("mod1");
+        Path dir2 = subdir("mod2");
 
-        var tui = new UpdatesTui(deps, pomPath, "g:a:1.0", (g, a) -> {
-            throw new IllegalStateException("network error");
-        });
+        Properties props1 = new Properties();
+        props1.setProperty("jackson.version", "2.17.0");
+        PilotProject p1 = createProject(
+                "com.example",
+                "mod1",
+                "1.0",
+                dir1,
+                List.of(dep("com.fasterxml", "jackson-core", "2.17.0")),
+                List.of(),
+                List.of(dep("com.fasterxml", "jackson-core", "${jackson.version}")),
+                List.of(),
+                props1);
 
-        var executor = Executors.newSingleThreadExecutor();
-        try {
-            tui.fetchUpdates(Runnable::run, executor).join();
-        } finally {
-            executor.shutdownNow();
+        Properties props2 = new Properties();
+        props2.setProperty("slf4j.version", "2.0.9");
+        PilotProject p2 = createProject(
+                "com.example",
+                "mod2",
+                "1.0",
+                dir2,
+                List.of(dep("org.slf4j", "slf4j-api", "2.0.9")),
+                List.of(),
+                List.of(dep("org.slf4j", "slf4j-api", "${slf4j.version}")),
+                List.of(),
+                props2);
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(p1, p2));
+
+        var tui = createTui(result, List.of(p1, p2));
+
+        for (var dep : result.allDependencies) {
+            dep.newestVersion = "3.0";
+            dep.updateType = VersionComparator.UpdateType.MAJOR;
+        }
+        for (var group : result.propertyGroups) {
+            group.newestVersion = "3.0";
+            group.updateType = VersionComparator.UpdateType.MAJOR;
         }
 
-        assertThat(tui.failedCount).isEqualTo(1);
-        assertThat(tui.loadedCount).isEqualTo(1);
-        assertThat(tui.loading).isFalse();
-        assertThat(tui.status).contains("1 lookup(s) failed");
+        tui.buildDisplayRows();
+
+        assertThat(tui.duplicatePropertyNames).isEmpty();
     }
 
     @Test
-    void fetchUpdatesMixedSuccessAndFailure() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "lib", "1.0", "compile", "jar"));
-        deps.add(new UpdatesTui.DependencyInfo("com.example", "broken", "1.0", "compile", "jar"));
+    void buildDisplayRowsCreatesGroupHeadersAndDeps() throws IOException {
+        Path dir = subdir("project");
 
-        var tui = new UpdatesTui(deps, pomPath, "g:a:1.0", (g, a) -> {
-            if ("broken".equals(a)) throw new IllegalStateException("timeout");
-            return List.of("2.0.0", "1.0.0");
-        });
+        Properties props = new Properties();
+        props.setProperty("jackson.version", "2.17.0");
+        PilotProject project = createProject(
+                "com.example",
+                "app",
+                "1.0",
+                dir,
+                List.of(
+                        dep("com.fasterxml", "jackson-core", "2.17.0"),
+                        dep("com.fasterxml", "jackson-databind", "2.17.0")),
+                List.of(),
+                List.of(
+                        dep("com.fasterxml", "jackson-core", "${jackson.version}"),
+                        dep("com.fasterxml", "jackson-databind", "${jackson.version}")),
+                List.of(),
+                props);
 
-        var executor = Executors.newSingleThreadExecutor();
-        try {
-            tui.fetchUpdates(Runnable::run, executor).join();
-        } finally {
-            executor.shutdownNow();
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+        assertThat(result.propertyGroups).hasSize(1);
+        assertThat(result.propertyGroups.get(0).dependencies).hasSize(2);
+
+        var tui = createTui(result, List.of(project));
+
+        for (var dep : result.allDependencies) {
+            dep.newestVersion = "2.18.0";
+            dep.updateType = VersionComparator.UpdateType.MINOR;
+        }
+        result.propertyGroups.get(0).newestVersion = "2.18.0";
+        result.propertyGroups.get(0).updateType = VersionComparator.UpdateType.MINOR;
+
+        tui.buildDisplayRows();
+
+        // 1 group header + 2 dependencies
+        assertThat(tui.displayRows).hasSize(3);
+        assertThat(tui.displayRows.get(0).isGroupHeader()).isTrue();
+        assertThat(tui.displayRows.get(0).propertyGroup.propertyName).isEqualTo("jackson.version");
+        assertThat(tui.displayRows.get(1).isGroupHeader()).isFalse();
+        assertThat(tui.displayRows.get(2).isGroupHeader()).isFalse();
+    }
+
+    @Test
+    void buildDisplayRowsEmptyWhenNoUpdates() {
+        PilotProject.Dep d = dep("com.example", "lib", "1.0");
+        PilotProject project =
+                createProject("com.example", "app", "1.0", tempDir, List.of(d), List.of(), List.of(d), List.of());
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+
+        var tui = createTui(result, List.of(project));
+        tui.buildDisplayRows();
+
+        assertThat(tui.displayRows).isEmpty();
+    }
+
+    @Test
+    void showDetailsDefaultsToTrue() {
+        PilotProject project = createProject("com.example", "app", "1.0", tempDir);
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+
+        var tui = createTui(result, List.of(project));
+
+        assertThat(tui.showDetails).isTrue();
+    }
+
+    @Test
+    void showDetailsCanBeToggledViaKeyHandler() {
+        PilotProject project = createProject("com.example", "app", "1.0", tempDir);
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+
+        var tui = createTui(result, List.of(project));
+
+        assertThat(tui.showDetails).isTrue();
+        tui.handleEvent(KeyEvent.ofChar('i'), null);
+        assertThat(tui.showDetails).isFalse();
+        tui.handleEvent(KeyEvent.ofChar('i'), null);
+        assertThat(tui.showDetails).isTrue();
+    }
+
+    @Test
+    void updateStatusIfDoneComputesPropertyGroupUpdates() throws IOException {
+        Path dir = subdir("project2");
+
+        Properties props = new Properties();
+        props.setProperty("jackson.version", "2.17.0");
+        PilotProject project = createProject(
+                "com.example",
+                "app",
+                "1.0",
+                dir,
+                List.of(dep("com.fasterxml", "jackson-core", "2.17.0")),
+                List.of(),
+                List.of(dep("com.fasterxml", "jackson-core", "${jackson.version}")),
+                List.of(),
+                props);
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+
+        var tui = createTui(result, List.of(project));
+
+        // Simulate version resolution for the dependency
+        var dep = result.allDependencies.get(0);
+        dep.newestVersion = "2.18.0";
+        dep.updateType = VersionComparator.UpdateType.MINOR;
+
+        // Simulate all dependencies loaded
+        tui.loadedCount = result.allDependencies.size() - 1;
+        tui.loading = true;
+
+        // Trigger updateStatusIfDone by incrementing loadedCount to match total
+        tui.loadedCount = result.allDependencies.size();
+        tui.updateStatusIfDone();
+
+        // After completion, loading should be false and group should have computed update
+        assertThat(tui.loading).isFalse();
+        assertThat(result.propertyGroups.get(0).newestVersion).isEqualTo("2.18.0");
+        assertThat(result.propertyGroups.get(0).updateType).isEqualTo(VersionComparator.UpdateType.MINOR);
+        assertThat(tui.status).contains("update");
+    }
+
+    @Test
+    void buildDisplayRowsWithFilterShowsOnlyMatchingUpdateType() throws IOException {
+        Path dir = subdir("project3");
+
+        PilotProject.Dep da = dep("com.example", "lib-a", "1.0");
+        PilotProject.Dep db = dep("com.example", "lib-b", "1.0");
+        PilotProject project =
+                createProject("com.example", "app", "1.0", dir, List.of(da, db), List.of(), List.of(da, db), List.of());
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+
+        var tui = createTui(result, List.of(project));
+
+        // Give lib-a a PATCH update
+        result.ungroupedDependencies.get(0).newestVersion = "1.0.1";
+        result.ungroupedDependencies.get(0).updateType = VersionComparator.UpdateType.PATCH;
+
+        // Give lib-b a MAJOR update
+        result.ungroupedDependencies.get(1).newestVersion = "2.0.0";
+        result.ungroupedDependencies.get(1).updateType = VersionComparator.UpdateType.MAJOR;
+
+        // Default filter: ALL - should show both
+        tui.buildDisplayRows();
+        assertThat(tui.displayRows).hasSize(2);
+
+        // Filter PATCH (f cycles ALL→PATCH) - only lib-a
+        tui.handleEvent(KeyEvent.ofChar('f'), null);
+        assertThat(tui.displayRows).hasSize(1);
+        assertThat(tui.displayRows.get(0).dependency.artifactId).isEqualTo("lib-a");
+
+        // Filter MAJOR (f cycles PATCH→MINOR→MAJOR) - only lib-b
+        tui.handleEvent(KeyEvent.ofChar('f'), null);
+        tui.handleEvent(KeyEvent.ofChar('f'), null);
+        assertThat(tui.displayRows).hasSize(1);
+        assertThat(tui.displayRows.get(0).dependency.artifactId).isEqualTo("lib-b");
+
+        // Filter ALL (f cycles MAJOR→ALL) - both again
+        tui.handleEvent(KeyEvent.ofChar('f'), null);
+        assertThat(tui.displayRows).hasSize(2);
+    }
+
+    @Test
+    void duplicatePropertyNamesFromSameOriginNotDetectedAsDuplicate() throws IOException {
+        Path dir = subdir("single");
+
+        Properties props = new Properties();
+        props.setProperty("jackson.version", "2.17.0");
+        PilotProject project = createProject(
+                "com.example",
+                "app",
+                "1.0",
+                dir,
+                List.of(
+                        dep("com.fasterxml", "jackson-core", "2.17.0"),
+                        dep("com.fasterxml", "jackson-databind", "2.17.0")),
+                List.of(),
+                List.of(
+                        dep("com.fasterxml", "jackson-core", "${jackson.version}"),
+                        dep("com.fasterxml", "jackson-databind", "${jackson.version}")),
+                List.of(),
+                props);
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+        // Same property from same origin -> only 1 group
+        assertThat(result.propertyGroups).hasSize(1);
+
+        var tui = createTui(result, List.of(project));
+
+        result.propertyGroups.get(0).newestVersion = "2.18.0";
+        result.propertyGroups.get(0).updateType = VersionComparator.UpdateType.MINOR;
+        for (var dep : result.allDependencies) {
+            dep.newestVersion = "2.18.0";
+            dep.updateType = VersionComparator.UpdateType.MINOR;
         }
 
-        assertThat(tui.loadedCount).isEqualTo(2);
-        assertThat(tui.failedCount).isEqualTo(1);
-        assertThat(tui.loading).isFalse();
-        assertThat(deps.get(0).newestVersion).isEqualTo("2.0.0");
-        assertThat(deps.get(1).newestVersion).isNull();
+        tui.buildDisplayRows();
+
+        // Same property name from same origin -> not a duplicate
+        assertThat(tui.duplicatePropertyNames).isEmpty();
     }
 
     @Test
-    void versionsNewestFirstReversesOrder() {
-        var result = UpdatesTui.versionsNewestFirst(List.of("1.0", "2.0", "3.0"));
-        assertThat(result).containsExactly("3.0", "2.0", "1.0");
+    void reactorRowIsGroupHeader() {
+        var group = new ReactorCollector.PropertyGroup("lib.version", "${lib.version}", "1.0", null);
+        var row = UpdatesTui.ReactorRow.group(group);
+        assertThat(row.isGroupHeader()).isTrue();
+        assertThat(row.propertyGroup).isSameAs(group);
+        assertThat(row.dependency).isNull();
     }
 
     @Test
-    void versionsNewestFirstEmptyList() {
-        var result = UpdatesTui.versionsNewestFirst(List.of());
-        assertThat(result).isEmpty();
+    void reactorRowIsDependency() {
+        var dep = new ReactorCollector.AggregatedDependency("g", "a");
+        var row = UpdatesTui.ReactorRow.dep(dep);
+        assertThat(row.isGroupHeader()).isFalse();
+        assertThat(row.dependency).isSameAs(dep);
+        assertThat(row.propertyGroup).isNull();
+    }
+
+    // -- Rendering tests (exercise code paths for coverage) --
+
+    @Test
+    void renderWithDetailPaneForPropertyGroup() throws IOException {
+        Path dir = subdir("render1");
+
+        Properties props = new Properties();
+        props.setProperty("jackson.version", "2.17.0");
+        PilotProject project = createProject(
+                "com.example",
+                "app",
+                "1.0",
+                dir,
+                List.of(
+                        dep("com.fasterxml", "jackson-core", "2.17.0"),
+                        dep("com.fasterxml", "jackson-databind", "2.17.0")),
+                List.of(),
+                List.of(
+                        dep("com.fasterxml", "jackson-core", "${jackson.version}"),
+                        dep("com.fasterxml", "jackson-databind", "${jackson.version}")),
+                List.of(),
+                props);
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+        var tui = createTui(result, List.of(project));
+
+        for (var dep : result.allDependencies) {
+            dep.newestVersion = "2.18.0";
+            dep.updateType = VersionComparator.UpdateType.MINOR;
+        }
+        result.propertyGroups.get(0).newestVersion = "2.18.0";
+        result.propertyGroups.get(0).updateType = VersionComparator.UpdateType.MINOR;
+
+        tui.loading = false;
+        tui.buildDisplayRows();
+
+        // Selected row 0 = property group header -> detail pane shows group info
+        renderFrame(tui);
     }
 
     @Test
-    void versionsNewestFirstConvertsToString() {
-        // Simulates Version objects that have toString()
-        var result = UpdatesTui.versionsNewestFirst(List.of(1, 2, 3));
-        assertThat(result).containsExactly("3", "2", "1");
+    void renderWithDetailPaneForDependency() throws IOException {
+        Path dir = subdir("render2");
+
+        Properties props = new Properties();
+        props.setProperty("jackson.version", "2.17.0");
+        PilotProject project = createProject(
+                "com.example",
+                "app",
+                "1.0",
+                dir,
+                List.of(dep("com.fasterxml", "jackson-core", "2.17.0")),
+                List.of(),
+                List.of(dep("com.fasterxml", "jackson-core", "${jackson.version}")),
+                List.of(),
+                props);
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+        var tui = createTui(result, List.of(project));
+
+        result.allDependencies.get(0).newestVersion = "2.18.0";
+        result.allDependencies.get(0).updateType = VersionComparator.UpdateType.MINOR;
+        result.propertyGroups.get(0).newestVersion = "2.18.0";
+        result.propertyGroups.get(0).updateType = VersionComparator.UpdateType.MINOR;
+
+        tui.loading = false;
+        tui.buildDisplayRows();
+
+        // Select row 1 = dependency under the group -> detail pane shows dep info
+        assertThat(tui.displayRows).hasSizeGreaterThanOrEqualTo(2);
+        // Move selection down to the dependency row via key handler
+        tui.handleEvent(KeyEvent.ofKey(KeyCode.DOWN), null);
+        renderFrame(tui);
     }
 
     @Test
-    void applyVersionResultAllPreviews() {
-        var deps = new ArrayList<UpdatesTui.DependencyInfo>();
-        var dep = new UpdatesTui.DependencyInfo("com.example", "lib", "1.0", "compile", "jar");
-        deps.add(dep);
+    void renderWithDetailPaneHidden() throws IOException {
+        Path dir = subdir("render3");
 
-        var tui = new UpdatesTui(deps, pomPath, "g:a:1.0", (g, a) -> List.of());
-        tui.applyVersionResult(dep, List.of("2.0.0-SNAPSHOT", "1.5.0-beta1", "1.2.0-alpha1"));
+        PilotProject.Dep d = dep("com.example", "lib", "1.0");
+        PilotProject project =
+                createProject("com.example", "app", "1.0", dir, List.of(d), List.of(), List.of(d), List.of());
 
-        assertThat(dep.newestVersion).isNull();
-        assertThat(dep.updateType).isNull();
-        assertThat(tui.loadedCount).isEqualTo(1);
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+        var tui = createTui(result, List.of(project));
+
+        result.ungroupedDependencies.get(0).newestVersion = "2.0";
+        result.ungroupedDependencies.get(0).updateType = VersionComparator.UpdateType.MAJOR;
+
+        tui.loading = false;
+        tui.showDetails = false;
+        tui.buildDisplayRows();
+
+        renderFrame(tui);
+    }
+
+    @Test
+    void renderWithDuplicatePropertyNames() throws IOException {
+        Path dir1 = subdir("render4a");
+        Path dir2 = subdir("render4b");
+
+        Properties props1 = new Properties();
+        props1.setProperty("lib.version", "1.0");
+        PilotProject p1 = createProject(
+                "com.example",
+                "bom-a",
+                "1.0",
+                dir1,
+                List.of(dep("com.example", "lib-a", "1.0")),
+                List.of(),
+                List.of(dep("com.example", "lib-a", "${lib.version}")),
+                List.of(),
+                props1);
+
+        Properties props2 = new Properties();
+        props2.setProperty("lib.version", "2.0");
+        PilotProject p2 = createProject(
+                "com.example",
+                "bom-b",
+                "1.0",
+                dir2,
+                List.of(dep("com.example", "lib-b", "2.0")),
+                List.of(),
+                List.of(dep("com.example", "lib-b", "${lib.version}")),
+                List.of(),
+                props2);
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(p1, p2));
+        var tui = createTui(result, List.of(p1, p2));
+
+        for (var dep : result.allDependencies) {
+            dep.newestVersion = "3.0";
+            dep.updateType = VersionComparator.UpdateType.MAJOR;
+        }
+        for (var group : result.propertyGroups) {
+            group.newestVersion = "3.0";
+            group.updateType = VersionComparator.UpdateType.MAJOR;
+        }
+
+        tui.loading = false;
+        tui.buildDisplayRows();
+
+        assertThat(tui.duplicatePropertyNames).containsExactly("lib.version");
+        renderFrame(tui);
+    }
+
+    @Test
+    void renderEmptyDepsView() {
+        PilotProject project = createProject("com.example", "app", "1.0", tempDir);
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+
+        var tui = createTui(result, List.of(project));
+        tui.loading = false;
+        tui.buildDisplayRows();
+
+        renderFrame(tui);
+    }
+
+    @Test
+    void renderLoadingState() {
+        PilotProject.Dep d = dep("com.example", "lib", "1.0");
+        PilotProject project =
+                createProject("com.example", "app", "1.0", tempDir, List.of(d), List.of(), List.of(d), List.of());
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+        var tui = createTui(result, List.of(project));
+        tui.loading = true;
+        tui.loadedCount = 0;
+
+        renderFrame(tui);
+    }
+
+    @Test
+    void renderWithUngroupedDependency() throws IOException {
+        Path dir = subdir("render5");
+
+        PilotProject.Dep d = dep("com.example", "lib", "1.0");
+        PilotProject project =
+                createProject("com.example", "app", "1.0", dir, List.of(d), List.of(), List.of(d), List.of());
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+        var tui = createTui(result, List.of(project));
+
+        result.ungroupedDependencies.get(0).newestVersion = "1.0.1";
+        result.ungroupedDependencies.get(0).updateType = VersionComparator.UpdateType.PATCH;
+
+        tui.loading = false;
+        tui.buildDisplayRows();
+
+        renderFrame(tui);
+    }
+
+    @Test
+    void renderWithSelectedUngroupedDependency() throws IOException {
+        Path dir = subdir("render6");
+
+        PilotProject.Dep d = dep("com.example", "lib", "1.0");
+        PilotProject project =
+                createProject("com.example", "app", "1.0", dir, List.of(d), List.of(), List.of(d), List.of());
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+        var tui = createTui(result, List.of(project));
+
+        var dep = result.ungroupedDependencies.get(0);
+        dep.newestVersion = "2.0.0";
+        dep.updateType = VersionComparator.UpdateType.MAJOR;
+        dep.selected = true;
+
+        tui.loading = false;
+        tui.buildDisplayRows();
+
+        // detail pane for ungrouped dependency (no property info)
+        renderFrame(tui);
+    }
+
+    @Test
+    void renderPropertyGroupWithoutUpdate() throws IOException {
+        Path dir = subdir("render7");
+
+        Properties props = new Properties();
+        props.setProperty("jackson.version", "2.17.0");
+        PilotProject project = createProject(
+                "com.example",
+                "app",
+                "1.0",
+                dir,
+                List.of(dep("com.fasterxml", "jackson-core", "2.17.0")),
+                List.of(),
+                List.of(dep("com.fasterxml", "jackson-core", "${jackson.version}")),
+                List.of(),
+                props);
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+        var tui = createTui(result, List.of(project));
+
+        // Dependency has update but group doesn't have a computed newest yet
+        result.allDependencies.get(0).newestVersion = "2.18.0";
+        result.allDependencies.get(0).updateType = VersionComparator.UpdateType.MINOR;
+        // Group has no newestVersion -> group.hasUpdate() is false but deps do
+
+        tui.loading = false;
+        tui.buildDisplayRows();
+
+        if (!tui.displayRows.isEmpty()) {
+            renderFrame(tui);
+        }
+    }
+
+    @Test
+    void renderWithManagedDependencyScopes() throws IOException {
+        Path dir = subdir("render8");
+
+        PilotProject.Dep d = dep("com.example", "lib", "1.0", "compile");
+        PilotProject project =
+                createProject("com.example", "app", "1.0", dir, List.of(d), List.of(), List.of(d), List.of());
+
+        ReactorCollector.CollectionResult result = ReactorCollector.collect(List.of(project));
+        var tui = createTui(result, List.of(project));
+
+        result.ungroupedDependencies.get(0).newestVersion = "1.1.0";
+        result.ungroupedDependencies.get(0).updateType = VersionComparator.UpdateType.MINOR;
+
+        tui.loading = false;
+        tui.buildDisplayRows();
+
+        renderFrame(tui);
     }
 }
